+++ src/constants/codepush-deployments.js
@@ -3,27 +3,32 @@
 module.exports = {
     Production: {
         key: 'LK2pdqNyAWGXr6-C-Ljk4kRTkgVJ4JUJX0auf',
-        title: 'Prod',
+        name: 'Production',
+        displayName: 'Prod',
         description: 'self-explanatory',
     },
     Staging: {
         key: '5yYuBL9RjePwiiuev-NzBhY1h-V-4JUJX0auf',
-        title: 'Staging',
+        name: 'Staging',
+        displayName: 'Staging',
         description: 'self-explanatory',
     },
     StagingBeta: {
         key: 'biT1JBowMsBtOYS8Rw8EzzHK6Pko4JUJX0auf',
-        title: 'Staging-beta',
+        name: 'StagingBeta',
+        displayName: 'Staging-beta',
         description: 'self-explanatory',
     },
     LocalTest: {
         key: 'JJGyWDPM2J2pfSNf3vwim1fERlvp4JUJX0auf',
-        title: 'Local1',
+        name: 'LocalTest',
+        displayName: 'Local1',
         description: 'use this to test the CodePush process *only* on your local machine',
     },
     LocalTest2: {
         key: 'rlE2Ld9i_ZOecfyCwqYek_CQxPqd4JUJX0auf',
-        title: 'Local2',
+        name: 'LocalTest2',
+        displayName: 'Local2',
         description: 'another (temp) local test deployment',
     },
 };